pyproject.toml

# ==== Project metadata ====

[project]
name = "validibot"
version = "0.1.0"
description = "A simple data validation service"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "Proprietary" }
authors = [{ name = "Daniel McQuillen" }]
dependencies = [
  "argon2-cffi==25.1.0",
  "authlib==1.6.5",
  "bleach==6.3.0",
  "cel-python==0.4.0",
  "crispy-bootstrap5==2025.6",
  "cryptography==46.0.3",
  "defusedxml==0.7.1",
  "django==6.0",
  "django-allauth[mfa]==65.13.1",
  "django-anymail[postmark]==13.1",
  "django-cloud-tasks>=0.0.4",
  "django-cors-headers==4.9.0",
  "django-crispy-forms==2.5",
  "django-environ==0.12.0",
  "django-filter==25.2",
  "django-markdownify==0.9.6",
  "django-model-utils==5.0.0",
  "django-redis==6.0.0",
  "django-storages[google,s3]==1.14.6",
  "djangorestframework==3.16.1",
  "drf-spectacular==0.29.0",
  "fmpy==0.3.26",
  "google-cloud-kms>=3.7.0",
  "google-cloud-tasks>=2.20.0",
  "gunicorn==23.0.0",
  "hiredis==3.3.0",
  "jsonschema==4.25.1",
  "lxml==6.0.2",
  "markdown2==2.5.4",
  "nh3==0.3.2",
  "pillow==12.0.0",
  "psycopg[c]==3.3.2",
  "pydantic==2.12.5",
  "python-slugify==8.0.4",
  "redis==7.1.0",
  "sentry-sdk==2.47.0",
  # vb-shared: Our shared library for Validibot integrations (e.g., EnergyPlus).
  # This Git URL is used when building Docker images (local or production).
  # For local non-Docker development, see [tool.uv.sources] below which overrides
  # this to use the local vb_shared_dev folder instead.
  "vb-shared @ git+https://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:x-oauth-basic@github.com/danielmcquillen/vb_shared.git@main",
  "uvicorn-worker==0.4.0",
  "uvicorn[standard]==0.38.0",
  "whitenoise==6.11.0",
  "google-cloud-run>=0.12.0",
]

# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = ["tests.py", "test_*.py"]
testpaths = [
  "tests",
  "validibot",
] # exclude sibling dev repos (sv_modal_dev, vb_shared_dev) from discovery
markers = [
  "integration: marks tests as integration (may require platform-specific binaries)",
]

# ignore test_use_cases
norecursedirs = [
  "sv_modal_dev",
  "vb_shared_dev",
  "vb_validators_dev",
  "tests/tests_integration", # We run these manually
]

# ==== Coverage ====
[tool.coverage.run]
include = ["validibot/**"]
# We should run tests on sv_modal and vb_shared separately.
omit = ["*/migrations/*", "*/tests/*", "sv_modal_dev", "vb_shared_dev", "vb_validators_dev"]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.ruff]
target-version = "py312"
# Exclude a variety of commonly ignored directories.
extend-exclude = ["*/migrations/*.py", "staticfiles/*", ]

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "C90",
  "I",
  "N",
  "UP",
  "YTT",
  # "ANN", # flake8-annotations: we should support this in the future but 100+ errors atm
  "ASYNC",
  "S",
  "BLE",
  "FBT",
  "B",
  "A",
  "COM",
  "C4",
  "DTZ",
  "T10",
  "DJ",
  "EM",
  "EXE",
  "FA",
  'ISC',
  "ICN",
  "G",
  'INP',
  'PIE',
  "T20",
  'PYI',
  'PT',
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SLOT",
  "SIM",
  "TID",
  "TC",
  "INT",
  # "ARG", # Unused function argument
  "PTH",
  "ERA",
  "PD",
  "PGH",
  "PL",
  "TRY",
  "FLY",
  # "NPY",
  # "AIR",
  "PERF",
  # "FURB",
  # "LOG",
  "RUF",
]
ignore = [
  "S101",   # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "SIM102", # sometimes it's better to nest
  # of types for comparison.
  # Deactivated because it can make the code slow:
  # https://github.com/astral-sh/ruff/issues/7871
  "TRY003",  # Use of bare except
  "EM102",
  "PT009",   # don't comment about tests using assertEqual etc
  "RET504",  #Unnecessary assignment to `filename` before `return` statement
  "ERA001",  # We want to put code in comments
  "C901",    # 'Function is too complex' ... I want to have some functions that have more than 6 args.
  "PLR0913", # Function has too many arguments (14/10)
  "BLE001",  #We want blind exceptions
  "PLR0912", # Too many branches (13/12)
  "PLR0911", # Too many statements (50/50)
  "EM101",   # Allow use of 'print' function
  "W293",    # blank line contains whitespace
  "W291",    # trailing whitespace
  "PLR0915", # Too many statements in a class (112/100)
  "RUF059",  # Allow use of dataclass with mutable default values
  "PLC0415", # Allow imports in places other than top of file
  "C416",    # Allow Unnecessary dict comprehension (rewrite using `dict()`)
  "COM812",  # Allow Trailing comma missing
  "PERF401", # Allow use of list comprehensions that could be replaced with generator expressions
  "G004",    # Allow f-strings in logging statements
]

[tool.ruff.lint.isort]
force-single-line = true

# ==== uv source overrides (local development only) ====
# When developing locally WITHOUT Docker, uv uses this section to override
# the vb-shared dependency. Instead of cloning from Git, it loads the package
# from the vb_shared_dev folder (a symlink to ../vb_shared). This allows us to
# edit vb-shared source code directly while working on Validibot.
#
# When building Docker images, we pass `--no-sources` to uv sync, which skips
# this override and installs vb-shared from the Git URL in [project.dependencies].
[tool.uv.sources]
vb-shared = { path = "vb_shared_dev", editable = true }

[dependency-groups]
dev = [
  # Testing
  "coverage==7.12.0",
  "django-coverage-plugin==3.2.0",
  "factory-boy==3.3.3",
  "pytest==9.0.1",
  "pytest-django==4.11.1",
  "pytest-sugar==1.1.1",
  "selenium==4.38.0",
  # Linting and formatting
  "djlint==1.36.4",
  "pre-commit==4.5.0",
  "ruff==0.14.6",
  # Type checking
  "django-stubs[compatible-mypy]==5.2.8",
  "djangorestframework-stubs==3.16.6",
  "mypy==1.19.0",
  # Debugging and development utilities
  "django-debug-toolbar==6.1.0",
  "django-extensions==4.1",
  "ipdb==0.13.13",
  "watchfiles==1.1.1",
  "werkzeug[watchdog]==3.1.4",
  # Documentation
  "mkdocs==1.6.1",
  "mkdocs-material==9.7.0",
  "mkdocs-print-site-plugin==2.8",
  # Local vb-shared development (uses tool.uv.sources override)
  "vb-shared",
]
